Move event::Handler methods to Widget; remove Handler
