--- crates/kas-core/src/event/handler.rs
@@ -1,86 +1,0 @@
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License in the LICENSE-APACHE file or at:
-//     https://www.apache.org/licenses/LICENSE-2.0
-
-//! Event handling - handler
-
-use super::*;
-use crate::WidgetConfig;
-#[allow(unused)]
-use crate::{Layout, Widget}; // for doc-links
-use kas_macros::autoimpl;
-
-/// Event handling for a [`Widget`]
-///
-/// This trait is part of the [`Widget`] family and is derived by
-/// [`derive(Widget)`] unless `#[handler(handle = noauto)]`
-/// or `#[handler(noauto)]` is used.
-///
-/// [`derive(Widget)`]: ../macros/index.html#the-derivewidget-macro
-#[autoimpl(for<T: trait + ?Sized> Box<T>)]
-pub trait Handler: WidgetConfig {
-    /// Handle an event sent to this widget
-    ///
-    /// An [`Event`] is some form of user input, timer or notification.
-    ///
-    /// This is the primary event handler for a widget. Secondary handlers are:
-    ///
-    /// -   If this method returns [`Response::Unused`], then
-    ///     [`Handler::handle_unused`] is called on each parent until the event
-    ///     is used (or the root widget is reached)
-    /// -   If a message is left on the stack by [`EventMgr::push_msg`], then
-    ///     [`Handler::handle_message`] is called on each parent until the stack is
-    ///     empty (failing to empty the stack results in a warning in the log).
-    /// -   If any scroll state is set by [`EventMgr::set_scroll`], then
-    ///     [`Handler::handle_scroll`] is called for each parent
-    ///
-    /// Default implementation: do nothing; return [`Response::Unused`].
-    #[inline]
-    fn handle_event(&mut self, mgr: &mut EventMgr, event: Event) -> Response {
-        let _ = (mgr, event);
-        Response::Unused
-    }
-
-    /// Handle an event sent to child `index` but left unhandled
-    ///
-    /// Default implementation: call [`Self::handle_event`] with `event`.
-    #[inline]
-    fn handle_unused(&mut self, mgr: &mut EventMgr, index: usize, event: Event) -> Response {
-        let _ = index;
-        self.handle_event(mgr, event)
-    }
-
-    /// Handler for messages from children/descendants
-    ///
-    /// This method is called when a child leaves a message on the stack. *Some*
-    /// parent or ancestor widget should read this message.
-    ///
-    /// The default implementation does nothing.
-    #[inline]
-    fn handle_message(&mut self, mgr: &mut EventMgr, index: usize) {
-        let _ = (mgr, index);
-    }
-
-    /// Handler for scrolling
-    ///
-    /// When a child calls [`EventMgr::set_scroll`] with a value other than
-    /// [`Scroll::None`], this method is called. (This method is not called
-    /// after [`Self::handle_event`] or other handlers called on self.)
-    ///
-    /// Note that [`Scroll::Rect`] values are in the child's coordinate space,
-    /// and must be translated to the widget's own coordinate space by this
-    /// method (this is not done by the default implementation since any widget
-    /// with non-zero translation very likely wants to implement this method
-    /// anyway).
-    ///
-    /// If the child is in an independent coordinate space, then this method
-    /// should call `mgr.set_scroll(Scroll::None)` to avoid any reactions to
-    /// child's scroll requests.
-    ///
-    /// The default implementation does nothing.
-    #[inline]
-    fn handle_scroll(&mut self, mgr: &mut EventMgr, scroll: Scroll) {
-        let _ = (mgr, scroll);
-    }
-}
