+++ crates/kas-core/src/event/manager/mgr_pub.rs
@@ -527,7 +527,7 @@ impl<'a> EventMgr<'a> {
     /// When setting [`Scroll::Rect`], use the widgets own coordinate space.
     ///
     /// Note that calling this method has no effect on the widget itself, but
-    /// affects parents via their [`Handler::handle_scroll`] method.
+    /// affects parents via their [`Widget::handle_scroll`] method.
     #[inline]
     pub fn set_scroll(&mut self, scroll: Scroll) {
         self.scroll = scroll;
