+++ crates/kas-widgets/src/adapter/map.rs
@@ -45,7 +45,7 @@ impl_scope! {
         }
     }
 
-    impl Handler for Self {
+    impl Widget for Self {
         fn handle_message(&mut self, mgr: &mut EventMgr, _: usize) {
             if let Some(msg) = mgr.try_pop_msg() {
                 mgr.push_msg((self.map)(msg));
