+++ crates/kas-widgets/src/dialog.rs
@@ -48,7 +48,7 @@ impl_scope! {
         }
     }
 
-    impl kas::event::Handler for Self {
+    impl kas::Widget for Self {
         fn handle_message(&mut self, mgr: &mut EventMgr, _: usize) {
             if let Some(MessageBoxOk) = mgr.try_pop_msg() {
                 mgr.send_action(TkAction::CLOSE);
