+++ crates/kas-widgets/src/edit_field.rs
@@ -438,10 +438,7 @@ impl_scope! {
         }
     }
 
-    impl event::Handler for Self
-    where
-        G: 'static,
-    {
+    impl Widget for Self {
         fn handle_event(&mut self, mgr: &mut EventMgr, event: Event) -> Response {
             fn request_focus<G: EditGuard + 'static>(s: &mut EditField<G>, mgr: &mut EventMgr) {
                 if !s.has_key_focus && mgr.request_char_focus(s.id()) {
