+++ crates/kas-widgets/src/slider.rs
@@ -261,7 +261,7 @@ impl_scope! {
         }
     }
 
-    impl Handler for Self {
+    impl Widget for Self {
         fn handle_event(&mut self, mgr: &mut EventMgr, event: Event) -> Response {
             match event {
                 Event::NavFocus(true) => {
