+++ crates/kas-widgets/src/splitter.rs
@@ -263,7 +263,7 @@ impl_scope! {
         }
     }
 
-    impl Handler for Self {
+    impl Widget for Self {
         fn handle_message(&mut self, mgr: &mut EventMgr, index: usize) {
             if (index & 1) == 1 {
                 if let Some(MsgPressFocus) = mgr.try_pop_msg() {
