+++ examples/data-list-view.rs
@@ -220,7 +220,7 @@ fn main() -> kas::shell::Result<()> {
             #[widget] _ = TextButton::new_msg("↓↑", Control::Dir),
             n: usize = 3,
         }
-        impl Handler for Self {
+        impl Widget for Self {
             fn handle_message(&mut self, mgr: &mut EventMgr, index: usize) {
                 if index == widget_index![self.edit] {
                     if let Some(n) = mgr.try_pop_msg::<usize>() {
@@ -265,7 +265,7 @@ fn main() -> kas::shell::Result<()> {
                 #[widget] list: ScrollBars<MyList> =
                     ScrollBars::new(list).with_bars(false, true),
             }
-            impl Handler for Self {
+            impl Widget for Self {
                 fn handle_message(&mut self, mgr: &mut EventMgr, _: usize) {
                     if let Some(control) = mgr.try_pop_msg::<Control>() {
                         match control {
