+++ examples/stopwatch.rs
@@ -8,11 +8,11 @@
 use std::time::{Duration, Instant};
 
 use kas::class::HasString;
-use kas::event::{Event, EventMgr, Handler, Response};
+use kas::event::{Event, EventMgr, Response};
 use kas::layout::SetRectMgr;
 use kas::macros::make_widget;
 use kas::widgets::{Frame, Label, TextButton, Window};
-use kas::WidgetExt;
+use kas::{Widget, WidgetExt};
 
 #[derive(Clone, Debug)]
 struct MsgReset;
@@ -39,7 +39,7 @@ fn make_window() -> Box<dyn kas::Window> {
                 mgr.enable_alt_bypass(self.id_ref(), true);
             }
         }
-        impl Handler for Self {
+        impl Widget for Self {
             fn handle_event(&mut self, mgr: &mut EventMgr, event: Event) -> Response {
                 match event {
                     Event::TimerUpdate(0) => {
